refactor(api): add meta for fn call, optimize code
