+++ packages/api/src/executor/index.ts
@@ -4,4 +4,4 @@ export * from './ConstantExecutor';
 export * from './StorageQueryExecutor';
 export * from './ErrorExecutor';
 export * from './EventExecutor';
-export * from './CallExecutor';
+export * from './RuntimeCallExecutor';
